fix: use new apis for trigger/input component from ember-power-select
