+++ packages/core/addon/components/eui-combo-box/trigger/input/index.hbs
@@ -1,0 +1,38 @@
+{{!template-lint-disable}}
+{{#if (and this.maybePlaceholder (not @select.searchText))}}
+  <p class="euiComboBoxPlaceholder">
+    {{this.maybePlaceholder}}
+  </p>
+{{/if}}
+<div
+  class="euiComboBox__input"
+  style="font-size: 14px; display: inline-block; position: relative;"
+>
+  <input
+    tabindex="-1"
+    style="opacity: 0px; width:0px; height:0px; position: absolute; top: 40%; border:solid 1px transparent !important; margin:0px !important;"
+    class="fake-input-for-html-form-validity"
+    {{validatable-control @isInvalid}}
+  />
+  <input
+    class="ember-power-select-trigger-multiple-input euiComboBox__input"
+    autocomplete="off"
+    autocorrect="off"
+    autocapitalize="off"
+    autofocus={{@autoFocus}}
+    spellcheck={{false}}
+    id="ember-power-select-trigger-multiple-input-{{@select.uniqueId}}"
+    value={{@select.searchText}}
+    aria-controls={{@listboxId}}
+    style={{this.triggerMultipleInputStyle}}
+    disabled={{@select.disabled}}
+    tabindex={{@tabindex}}
+    form="power-select-fake-form"
+    {{on "focus" @onFocus}}
+    {{on "blur" @onBlur}}
+    {{on "input" this.handleInput}}
+    {{on "keydown" this.handleKeydown}}
+    {{did-insert this.storeInputStyles}}
+  />
+</div>
+{{!template-lint-enable}}
