+++ packages/core/app/components/eui-combo-box/trigger/input/index.js
@@ -1,0 +1,1 @@
+export { default } from '@ember-eui/core/components/eui-combo-box/trigger/input';
